Update and reorganize intermediate results of vignettes
Because of computational reasons some cells of the vignettes are not rerun each time. They rely on intermediately stored files. While this is error-prone, it had been chosen as a computationally pragmatic way.

Here, we update the results and streamline the generation of the result files. The aim is to make more evident where each result was generated and is used.

Lastly the sensitivity vignette has been changed to calibration of kphio, kphio_par_b, and kc_jmax; instead of kphio, kc_jmax, and soilm_betao.