Update tests, vignette, and documentation for gw change
Previous commits introduced a new input parameter gw_calib.

To respect the testing with continuous integration any change should include updating of the
- vignettes/*.Rmd
- tests/testthat/*.R
- and documentation
- data/*.rda

Ideally this would be done in the same commit as the change to the model itself.

Note that steps 1 and 2 can be tested locally with devtools::build_vignettes() and with devtools::test()