Add tolerance to BiomeE regression tests
The selected reference values were generated on Ubuntu. The GitHub CI showed that some differences exist when comparing on macOS or on Windows. Most mean relative differences are tiny. I.e. below 1e-5. Rauto was frequently between 1e-4 and 1e-5.

Outlier was plantN on Windows with Mean relative differences of: 0.0005171742. Therefore we set the tolerance to 1e-3