Rename 'state' into 'steering_state'
This clarifies the term `state` in this context as referring to the steering variables of the current year. In another context `state` is already used in rsofun for physical state variables of the dynamical system. 

Essentially `steering_state` contains the steering object, i.e. the state of the simulation step, controlling if output is written etc.

This also clarifies the `steering` parameters by renaming them `steering_params`, reflecting the fact that these input parameters are constant throughout the whole simulation.