Introduce a new input parameter gw_calib (#287)
* update

* update

* update

* Homogenize whitespace

* Revert removal of data files

* Update tests, vignette, and documentation for gw change

Previous commits introduced a new input parameter gw_calib.

To respect the testing with continuous integration any change should include updating of the
- vignettes/*.Rmd
- tests/testthat/*.R
- and documentation
- data/*.rda

Ideally this would be done in the same commit as the change to the model itself.

Note that steps 1 and 2 can be tested locally with devtools::build_vignettes() and with devtools::test()

* Fix phydro bug by requiring `gw_calib`

* Use single Makevars as in branches 'main' and 'luluc'

* Remove daily and night rad output

* Remove daily and night rad output from tests

* Reduce whitespace changes by this PR

* Add gw_calib also to vignettes

* Fix gw_calib in examples of documentation

* Include gw_calib in sensitivity analysis

---------

Co-authored-by: Fabian Bernhard <10245680+fabern@users.noreply.github.com>